@@ -13,22 +13,21 @@
     "format": "prettier \"./src/**/*.{ts,tsx}\" --write",
     "new:page": "plop --plopfile ./generators/plopfile.js page",
     "new:component": "plop --plopfile ./generators/plopfile.js component",
-    "new:hook": "plop --plopfile ./generators/plopfile.js hook",
     "prepare": "husky install",
     "storybook": "start-storybook -p 6006 -s public",
     "build-storybook": "build-storybook -s public"
   },
   "dependencies": {
-    "axios": "^0.26.1",
-    "date-fns": "^2.28.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-icons": "^4.3.1",
-    "react-router-dom": "^6.2.2",
-    "react-scripts": "5.0.0",
-    "styled-components": "^5.3.3",
+    "axios": "^1.3.6",
+    "date-fns": "^2.29.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.8.0",
+    "react-router-dom": "^6.10.0",
+    "react-scripts": "5.0.1",
+    "styled-components": "^5.3.10",
     "typescript": "^4.6.2",
-    "web-vitals": "^2.1.4"
+    "web-vitals": "^3.3.1"
   },
   "devDependencies": {
     "@storybook/addon-actions": "^6.4.19",
@@ -43,10 +42,10 @@
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.4.1",
     "@types/node": "^17.0.21",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.13",
+    "@types/react": "^18.2.0",
+    "@types/react-dom": "^18.2.0",
     "@types/react-router-dom": "^5.3.3",
-    "@types/styled-components": "^5.1.24",
+    "@types/styled-components": "^5.1.26",
     "@typescript-eslint/eslint-plugin": "^5.14.0",
     "@typescript-eslint/parser": "^5.14.0",
     "babel-eslint": "^10.1.0",
@@ -61,8 +60,8 @@
     "eslint-plugin-react-hooks": "^4.3.0",
     "eslint-plugin-security": "^1.4.0",
     "husky": "^7.0.4",
-    "plop": "^3.0.5",
-    "prettier": "^2.5.1",
+    "plop": "^3.1.2",
+    "prettier": "^2.8.8",
     "source-map-explorer": "^2.5.2",
     "storybook-addon-outline": "^1.4.2"
   },
