@@ -1,7 +1,6 @@
-const { page, component, hook } = require('./settings');
+const { page, component } = require('./settings');
 
 module.exports = function (plop) {
   plop.setGenerator('page', page);
   plop.setGenerator('component', component);
-  plop.setGenerator('hook', hook);
 };
